Reorganize memory module
