--- apps/python/ACLIB/source/memory/ac_info.py
@@ -1,7 +1,0 @@
-import ac
-
-def active_car_only(func):
-    if ac.getFocusedCar() == 0:
-        return func()
-    else:
-        return - 1
